app/requirements.txt
fastapi
async-fastapi-jwt-auth
async-fastapi-jwt-auth[asymmetric]
sqlalchemy
python-dotenv